backend/requirements.txt
pandas==2.2.3
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
ta==0.11.0
beautifulsoup4==4.12.3
pykrx==1.0.48
django-model-utils==5.0.0
django-cors-headers==4.6.0
fake-useragent==0.1.13
finance-datareader==0.9.94
discord.py==2.4.0
python-telegram-bot==21.6
django-apscheduler==0.7.0
plotly==5.24.1
django-environ==0.11.2
django-extensions==3.2.3
IPython==8.31.0
djangorestframework-simplejwt==5.3.0
xmltodict==0.14.2
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
